feat: update app search
